Restructure the code base
