+++ actionweaver/__init__.py
@@ -1,2 +1,2 @@
 from .action import Action, ActionException, action
-from .mixin import ActionHandlerMixin
+from .mixins import ActionHandlerMixin
