+++ tests/utils/test_pydantic_utils.py
@@ -5,8 +5,8 @@ import unittest
 
 from pydantic import BaseModel, create_model
 
-from actionweaver.cache import cache, lru_cache, preserve_original_signature
-from actionweaver.utils import create_pydantic_model_from_func
+from actionweaver.utils.cache import cache, lru_cache, preserve_original_signature
+from actionweaver.utils.pydantic_utils import create_pydantic_model_from_func
 
 
 class UtilsTestCase(unittest.TestCase):
